# remove an unnecessary operation in `getLocationTree` for better LSP performance