The following are benchmarks of a command that executes `textDocument/documentHighlight` five times:

```
// Before this PR
Benchmark 1: neut foo
  Time (mean ± σ):     480.3 ms ±   6.0 ms    [User: 390.0 ms, System: 196.8 ms]
  Range (min … max):   471.9 ms … 489.5 ms    10 runs

↓

// After this PR
Benchmark 1: neut foo
  Time (mean ± σ):      47.5 ms ±   1.5 ms    [User: 17.1 ms, System: 18.9 ms]
  Range (min … max):    44.2 ms …  50.8 ms    57 runs
```